fix(delete): preserve parent type for non-empty subtree deletes
Rehabilitation of PR #732 (fixes #686) for the GROVE_V4 boundary:

- restructure delete_internal_on_transaction into the standard versioned
  dispatch (delete_internal_on_transaction/{mod.rs,v0.rs,v1.rs}), following
  insert/add_element_on_transaction
- v1 (GROVE_V4+) reuses the already-open parent Merk when deleting a
  non-empty child tree instead of reopening the parent layer labeled with
  the child's tree type, so delete propagation hashes and aggregates with
  the parent tree type
- v0 keeps the legacy reopen byte-for-byte for GROVE_V1..V3 (live in
  production; replay compatibility)
- regressions: CountTree / CountSumTree parents (aggregates settle),
  ProvableCountTree parent (link-hash binding stays verifiable under v4,
  corruption pinned under v3), Provable* child under plain parent
  (hash_for_link panic fixed under v4, panic pinned under v3), plain
  legacy path pinned under v3, gate values asserted in grovedb-version
