+++ grovedb-version/src/tests.rs
@@ -492,6 +492,30 @@ fn terminal_keys_is_legacy_until_v4() {
 }
 
 #[test]
+fn delete_internal_on_transaction_is_legacy_until_v4() {
+    // Reusing the already-open parent Merk for non-empty child tree deletes
+    // (issue #686) activates at GROVE_V4; v1-v3 are live in production and
+    // must keep the legacy reopen labeled with the child's tree type.
+    for v in [&GROVE_V1, &GROVE_V2, &GROVE_V3] {
+        assert_eq!(
+            v.grovedb_versions
+                .operations
+                .delete
+                .delete_internal_on_transaction,
+            0
+        );
+    }
+    assert_eq!(
+        GROVE_V4
+            .grovedb_versions
+            .operations
+            .delete
+            .delete_internal_on_transaction,
+        1
+    );
+}
+
+#[test]
 fn v1_replication_all_zero() {
     let rep = &GROVE_V1.grovedb_versions.replication;
     assert_eq!(rep.get_subtrees_metadata, 0);
